requirements.txt
# Core Dependencies
numpy==1.24.3
opencv-python==4.8.1.78
pillow==10.0.0
pytesseract==0.3.10
pyautogui==0.9.54
pynput==1.7.6

# Windows specific
pywin32==306

# Deep Learning
torch==2.1.0
torchvision==0.16.0

# Utilities
pyyaml==6.0.1
tqdm==4.66.1
matplotlib==3.7.2
tensorboard==2.14.0

# Optional (for advanced features)
gymnasium==0.29.1  # RL environment standard
stable-baselines3==2.1.0  # Pre-built RL algorithms (optional)